% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{column_means_and_vars}
\alias{column_means_and_vars}
\title{Column means and variances}
\usage{
column_means_and_vars(M)
}
\arguments{
\item{M}{the matrix containing the data}
}
\description{
Column means and variances
}
