% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{dataset_from_str}
\alias{dataset_from_str}
\title{Data name string to dataset}
\usage{
dataset_from_str(dataset_name)
}
\arguments{
\item{dataset_name}{the name of the dataset to load}
}
\value{
the dataset
}
\description{
Useful to load a dataset into a variable.
}
