% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{group_dissimilarity}
\alias{group_dissimilarity}
\title{group_dissimilarity}
\usage{
group_dissimilarity(dissimilarity, f, groups)
}
\arguments{
\item{dissimilarity}{A dissimilarity between individuals}

\item{groups}{a group membership vector}

\item{weight}{the weights of the individuals}
}
\description{
group_dissimilarity
}
