% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{group_similarity}
\alias{group_similarity}
\title{Group dissimilarity}
\usage{
group_similarity(dissimilarity, weight, groups)
}
\arguments{
\item{dissimilarity}{A dissimilarity between individuals}

\item{weight}{the weights of the individuals}

\item{groups}{a group membership vector}
}
\description{
Group dissimilarity
}
