% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italian_legislator_17__1__D_final__unweighted}
\alias{italian_legislator_17__1__D_final__unweighted}
\title{italian_legislator_17__1__D_final__unweighted}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 664 rows and 664 columns.
}
\usage{
italian_legislator_17__1__D_final__unweighted
}
\description{
italian_legislator_17__1__D_final__unweighted
}
\keyword{datasets}
