% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italian_legislator_17__1__weighted_vote_disputedness}
\alias{italian_legislator_17__1__weighted_vote_disputedness}
\title{italian_legislator_17__1__weighted_vote_disputedness}
\format{
An object of class \code{numeric} of length 24669.
}
\usage{
italian_legislator_17__1__weighted_vote_disputedness
}
\description{
italian_legislator_17__1__weighted_vote_disputedness
}
\keyword{datasets}
