% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italian_legislator_17__2__D_final}
\alias{italian_legislator_17__2__D_final}
\title{italian_legislator_17 2 D_final}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 665 rows and 665 columns.
}
\usage{
italian_legislator_17__2__D_final
}
\description{
italian_legislator_17 2 D_final
}
\keyword{datasets}
