% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italian_legislator_17__2__D_final__weighted}
\alias{italian_legislator_17__2__D_final__weighted}
\title{italian_legislator_17__2__D_final__weighted}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 665 rows and 665 columns.
}
\usage{
italian_legislator_17__2__D_final__weighted
}
\description{
italian_legislator_17__2__D_final__weighted
}
\keyword{datasets}
