% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italian_legislator_17__2__scores_matrix}
\alias{italian_legislator_17__2__scores_matrix}
\title{italian_legislator_17 2 scores_matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 670 rows and 24930 columns.
}
\usage{
italian_legislator_17__2__scores_matrix
}
\description{
italian_legislator_17 2 scores_matrix
}
\keyword{datasets}
