% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italian_legislator_17__3__null_councilors}
\alias{italian_legislator_17__3__null_councilors}
\title{italian_legislator_17 3 null_councilors}
\format{
An object of class \code{character} of length 5.
}
\usage{
italian_legislator_17__3__null_councilors
}
\description{
italian_legislator_17 3 null_councilors
}
\keyword{datasets}
