% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italian_legislator_17__3__weight}
\alias{italian_legislator_17__3__weight}
\title{italian_legislator_17 3 weight}
\format{
An object of class \code{numeric} of length 665.
}
\usage{
italian_legislator_17__3__weight
}
\description{
italian_legislator_17 3 weight
}
\keyword{datasets}
