% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italian_legislator_18__3__null_votes}
\alias{italian_legislator_18__3__null_votes}
\title{italian_legislator_18__3__null_votes}
\format{
An object of class \code{character} of length 26.
}
\usage{
italian_legislator_18__3__null_votes
}
\description{
italian_legislator_18__3__null_votes
}
\keyword{datasets}
