% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{save_with_name}
\alias{save_with_name}
\title{Data name string to dataset}
\usage{
save_with_name(variable, new_name, compress = "gzip")
}
\arguments{
\item{variable}{the original variable}

\item{new_name}{name of the data}

\item{compress}{compression type see save}
}
\description{
Useful to load a dataset into a variable.
}
