% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swiss_legislator_49__1__D_final}
\alias{swiss_legislator_49__1__D_final}
\title{swiss_legislator_49 1 D_final}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 228 rows and 228 columns.
}
\usage{
swiss_legislator_49__1__D_final
}
\description{
swiss_legislator_49 1 D_final
}
\keyword{datasets}
