% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swiss_legislator_49__1__scores_matrix}
\alias{swiss_legislator_49__1__scores_matrix}
\title{swiss_legislator_49 1 scores_matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 228 rows and 4245 columns.
}
\usage{
swiss_legislator_49__1__scores_matrix
}
\description{
swiss_legislator_49 1 scores_matrix
}
\keyword{datasets}
