% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swiss_legislator_49__2__null_councilors}
\alias{swiss_legislator_49__2__null_councilors}
\title{swiss_legislator_49 2 null_councilors}
\format{
An object of class \code{character} of length 0.
}
\usage{
swiss_legislator_49__2__null_councilors
}
\description{
swiss_legislator_49 2 null_councilors
}
\keyword{datasets}
