% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swiss_legislator_50__3__weight}
\alias{swiss_legislator_50__3__weight}
\title{swiss_legislator_50__3__weight}
\format{
An object of class \code{numeric} of length 228.
}
\usage{
swiss_legislator_50__3__weight
}
\description{
swiss_legislator_50__3__weight
}
\keyword{datasets}
