% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swiss_legislator_50__3__weighted_vote_disputedness}
\alias{swiss_legislator_50__3__weighted_vote_disputedness}
\title{swiss_legislator_50__3__weighted_vote_disputedness}
\format{
An object of class \code{numeric} of length 4452.
}
\usage{
swiss_legislator_50__3__weighted_vote_disputedness
}
\description{
swiss_legislator_50__3__weighted_vote_disputedness
}
\keyword{datasets}
